refactor: move shared code into libs
