+++ libs/@guardian/identity-auth/src/token.ts
@@ -13,124 +13,30 @@ import type {
 } from './@types/OAuth';
 import type {
 	AccessToken,
-	AccessTokenClaims,
 	CustomClaims,
 	IDToken,
-	IDTokenClaims,
-	JWK,
-	JWKS,
-	JWTHeader,
-	JWTObject,
 	JWTPayload,
 	TokenResponse,
 	Tokens,
 } from './@types/Token';
 import {
 	base64UrlEncode,
-	base64UrlToString,
 	generateCodeChallenge,
 	generateCodeVerifier,
 	generateSha256Hash,
 	getRandomString,
-	stringToBuffer,
 } from './crypto';
 import { OAuthError } from './error';
-
-/**
- * @name isAccessTokenClaims
- * @description Type guard for AccessTokenClaims
- *
- * @param claims - the claims to check
- * @returns	{boolean} - true if claims is an AccessTokenClaims object
- */
-const isAccessTokenClaims = (claims: unknown): claims is AccessTokenClaims => {
-	const maybeClaims = claims as AccessTokenClaims;
-	return (
-		isNonNullable(maybeClaims) &&
-		maybeClaims.aud !== undefined &&
-		maybeClaims.auth_time !== undefined &&
-		maybeClaims.cid !== undefined &&
-		maybeClaims.email_validated !== undefined &&
-		maybeClaims.exp !== undefined &&
-		maybeClaims.iat !== undefined &&
-		maybeClaims.identity_username !== undefined &&
-		maybeClaims.iss !== undefined &&
-		maybeClaims.jti !== undefined &&
-		maybeClaims.legacy_identity_id !== undefined &&
-		maybeClaims.scp !== undefined &&
-		maybeClaims.sub !== undefined &&
-		maybeClaims.uid !== undefined &&
-		maybeClaims.ver !== undefined
-	);
-};
-
-/**
- * @name isAccessToken
- * @description Type guard for AccessToken
- *
- * @param token - the token to check
- * @returns	{boolean} - true if token is an AccessToken object
- */
-export const isAccessToken = (token: unknown): token is AccessToken => {
-	const maybeToken = token as AccessToken;
-	return (
-		isNonNullable(maybeToken) &&
-		maybeToken.accessToken !== undefined &&
-		maybeToken.expiresAt !== undefined &&
-		maybeToken.scopes !== undefined &&
-		maybeToken.tokenType !== undefined &&
-		isAccessTokenClaims(maybeToken.claims)
-	);
-};
-
-/**
- * @name isIDTokenClaims
- * @description Type guard for IDTokenClaims
- *
- * @param claims - the claims to check
- * @returns	{boolean} - true if claims is an IDTokenClaims object
- */
-const isIDTokenClaims = (claims: unknown): claims is IDTokenClaims => {
-	const maybeClaims = claims as IDTokenClaims;
-	return (
-		isNonNullable(maybeClaims) &&
-		maybeClaims.amr !== undefined &&
-		maybeClaims.at_hash !== undefined &&
-		maybeClaims.aud !== undefined &&
-		maybeClaims.auth_time !== undefined &&
-		maybeClaims.exp !== undefined &&
-		maybeClaims.iat !== undefined &&
-		maybeClaims.identity_username !== undefined &&
-		maybeClaims.idp !== undefined &&
-		maybeClaims.iss !== undefined &&
-		maybeClaims.jti !== undefined &&
-		maybeClaims.legacy_identity_id !== undefined &&
-		maybeClaims.name !== undefined &&
-		maybeClaims.nonce !== undefined &&
-		maybeClaims.preferred_username !== undefined &&
-		maybeClaims.sub !== undefined &&
-		maybeClaims.user_groups !== undefined &&
-		maybeClaims.ver !== undefined
-	);
-};
-
-/**
- * @name isIDToken
- * @description Type guard for IDToken
- *
- * @param token - the token to check
- * @returns	{boolean} - true if token is an IDToken object
- */
-export const isIDToken = (token: unknown): token is IDToken => {
-	const maybeToken = token as IDToken;
-	return (
-		isNonNullable(maybeToken) &&
-		maybeToken.expiresAt !== undefined &&
-		maybeToken.idToken !== undefined &&
-		maybeToken.scopes !== undefined &&
-		isIDTokenClaims(maybeToken.claims)
-	);
-};
+import {
+	decodeToken,
+	genericVerifyAccessTokenClaims,
+	genericVerifyIdTokenClaims,
+	isAccessToken,
+	isAccessTokenClaims,
+	isIDToken,
+	isIDTokenClaims,
+	verifySignature,
+} from './lib/token';
 
 /**
  * @name isOAuthAuthorizeResponseError
@@ -199,57 +105,7 @@ const calculateClockSkew = (now: number, token: string): number => {
 	return now - iat;
 };
 
-/**
- * @name decodeToken
- * @description Decodes a JWT token into its parts, i.e header, payload and signature
- *
- * @param token `string` - JWT token
- * @returns JWTObject - object containing the decoded token parts
- */
-export const decodeToken = <T extends CustomClaims = CustomClaims>(
-	token: string,
-): JWTObject<T> => {
-	// attempt to split the token into its parts
-	const [headerStr, payloadStr, signature] = token.split('.');
-
-	try {
-		// check that all parts are present
-		if (
-			headerStr === undefined ||
-			payloadStr === undefined ||
-			signature === undefined
-		) {
-			throw 'Missing token parts';
-		}
-
-		const header = JSON.parse(base64UrlToString(headerStr)) as JWTHeader;
-		const payload = JSON.parse(base64UrlToString(payloadStr)) as JWTPayload<T>;
-
-		// validate the header, all okta oauth tokens have an alg and kid claim
-		if (!header.alg || !header.kid) {
-			throw 'Missing algorithm in header';
-		}
-
-		// validate the payload, all okta oauth tokens have an jti claim
-		if (!payload.jti) {
-			throw 'Possibly malformed payload';
-		}
-
-		return {
-			header,
-			payload,
-			signature,
-		};
-	} catch (error) {
-		throw new OAuthError({
-			error: 'invalid_token',
-			error_description: 'Malformed token',
-			message: error as string,
-		});
-	}
-};
-
-interface DecodeTokenParams {
+interface DecodeTokensParams {
 	accessTokenRaw: string;
 	accessTokenClockSkew: number;
 	idTokenRaw: string;
@@ -278,7 +134,7 @@ const decodeTokens = <
 	idTokenClockSkew,
 	nonce,
 	options,
-}: DecodeTokenParams): Tokens<AC, IC> => {
+}: DecodeTokensParams): Tokens<AC, IC> => {
 	// get the current time in seconds
 	const now = Math.floor(Date.now() / 1000);
 
@@ -385,74 +241,6 @@ const decodeTokens = <
 };
 
 /**
- * @name verifyIdTokenClaims
- * @description Verifies the claims of an ID token are valid and expected
- *
- * @param decoded - decoded ID token
- * @param claims - ID token claims
- * @param options - RequiredIdentityAuthOptions
- *
- * @throws Error - if the claims are invalid
- * @returns void
- */
-export const verifyIdTokenClaims = (
-	decoded: IDToken,
-	claims: JWTPayload,
-	options: RequiredIdentityAuthOptions,
-): void => {
-	// get the local time in seconds
-	const localTime = Math.floor(Date.now() / 1000);
-
-	// calculate the normalised time, which takes into account clock skew
-	const normalisedTime = localTime - decoded.clockSkew;
-
-	// check the nonce is valid
-	if (claims.nonce !== decoded.nonce) {
-		throw new OAuthError({
-			error: 'invalid_token',
-			error_description: 'Invalid nonce in ID token',
-			message: 'Invalid nonce in ID token',
-		});
-	}
-
-	// check the issuer is the one configured in the options
-	if (claims.iss !== options.issuer) {
-		throw new OAuthError({
-			error: 'invalid_token',
-			error_description: 'Invalid issuer in ID token',
-			message: 'Invalid issuer in ID token',
-		});
-	}
-
-	// check the audience is the one configured in the options (clientId)
-	if (claims.aud !== options.clientId) {
-		throw new OAuthError({
-			error: 'invalid_token',
-			error_description: 'Invalid audience in ID token',
-			message: 'Invalid audience in ID token',
-		});
-	}
-
-	// check that the iat (issued at) claim is present
-	if (!claims.iat) {
-		throw new OAuthError({
-			error: 'invalid_token',
-			error_description: 'Missing iat claim in ID token',
-			message: 'Token does not contain required claims',
-		});
-	}
-
-	// check the token wasn't issued in the future
-	if (claims.iat > normalisedTime) {
-		throw new OAuthError({
-			error: 'invalid_token',
-			error_description: 'Token was issued in the future',
-			message: 'Token was issued in the future',
-		});
-	}
-};
-
-/**
  * @name verifyAccessTokenWithAtHash
  * @description Verifies the access token with the id token at_hash claim
  *
@@ -486,123 +274,6 @@ export const verifyAccessTokenWithAtHash = async (
 	}
 	// if they do match, the token is valid, return void
 	return;
-};
-
-/**
- * @name verifySignature
- * @description Verifies the signature of a JWT token
- *
- * Get's the public key from the JWKS endpoint, imports it and verifies the signature
- * of a given JWT token
- *
- * @param issuer
- * @param jwt
- * @param token
- */
-export const verifySignature = async (
-	jwksUri: string,
-	jwt: JWTObject,
-	token: string,
-): Promise<void> => {
-	// fetch the JWKS, okta returns standard cache-control headers, so caching is handled by the browser
-	const jwksResponse = await fetch(jwksUri);
-
-	// throw an error if the response isn't ok
-	if (!jwksResponse.ok) {
-		throw new OAuthError({
-			error: 'failed_request',
-			error_description: 'Failed to fetch JWKS',
-			message: 'Failed to fetch JWKS',
-		});
-	}
-
-	// parse the JWKS response
-	const jwks = (await jwksResponse.json()) as JWKS;
-
-	// validate the JWKS, first check that there are keys
-	if (!jwks.keys || jwks.keys.length === 0) {
-		throw new OAuthError({
-			error: 'invalid_jwks',
-			error_description: 'No keys found in JWKS',
-			message: 'No keys found in JWKS',
-		});
-	}
-
-	// next loop through the keys and check that they have the required properties
-	for (const key of jwks.keys) {
-		if (!key.kty || !key.n || !key.e || !key.alg || !key.use || !key.kid) {
-			throw new OAuthError({
-				error: 'invalid_jwks',
-				error_description: 'Invalid key in JWKS',
-				message: 'Invalid key in JWKS',
-			});
-		}
-	}
-
-	// find the key that matches the kid in the JWT header
-	const key = jwks.keys.find((k: JWK) => k.kid === jwt.header.kid);
-
-	// throw an error if no key was found
-	if (key === undefined) {
-		throw new OAuthError({
-			error: 'invalid_jwks',
-			error_description: 'No key found for token in JWKS',
-			message: 'No key found for token in JWKS',
-		});
-	}
-
-	// create the algorithm to use for the verification
-	const algorithm: HmacImportParams = {
-		name: 'RSASSA-PKCS1-v1_5',
-		hash: { name: 'SHA-256' },
-	};
-
-	// import the public key using the jwk format
-	const publicKey = await window.crypto.subtle.importKey(
-		'jwk',
-		{
-			kty: key.kty,
-			n: key.n,
-			e: key.e,
-			alg: key.alg,
-			use: key.use,
-		},
-		algorithm,
-		true,
-		['verify'],
-	);
-
-	// split the token into it's parts
-	const [header, payload, sig] = token.split('.');
-
-	// throw an error if the token is malformed
-	if (header === undefined || payload === undefined || sig === undefined) {
-		throw new OAuthError({
-			error: 'invalid_token',
-			error_description: 'Malformed token',
-			message: 'Malformed token',
-		});
-	}
-
-	// create the data to verify, i.e the header and payload
-	const data = `${header}.${payload}`;
-
-	// verify the signature
-	const isValid = await window.crypto.subtle.verify(
-		algorithm,
-		publicKey,
-		stringToBuffer(base64UrlToString(sig)),
-		stringToBuffer(data),
-	);
-
-	// throw an error if the signature is invalid
-	if (!isValid) {
-		throw new OAuthError({
-			error: 'invalid_token',
-			error_description: 'Invalid signature',
-			message: 'Invalid signature',
-		});
-	}
 };
 
 /**
@@ -822,7 +493,13 @@ export const verifyTokens = async (
 		const idTokenJWT = decodeToken(idToken.idToken);
 
 		// verify the claims
-		verifyIdTokenClaims(idToken, idTokenJWT.payload, options);
+		genericVerifyIdTokenClaims({
+			claims: idTokenJWT.payload,
+			issuer: options.issuer,
+			clientId: options.clientId,
+			nonce: idToken.nonce,
+			clockSkew: idToken.clockSkew,
+		});
 
 		// verify the signature
 		await verifySignature(oauthUrls.keysUrl, idTokenJWT, idToken.idToken);
@@ -835,8 +512,14 @@ export const verifyTokens = async (
 
 		const accessTokenJWT = decodeToken(accessToken.accessToken);
 
-		// verify access token timestamps
-		verifyAccessTokenTimestamps(accessToken, accessTokenJWT.payload);
+		// verify access token claims
+		genericVerifyAccessTokenClaims({
+			claims: accessTokenJWT.payload,
+			audience: 'todo',
+			issuer: options.issuer,
+			clientId: options.clientId,
+			clockSkew: accessToken.clockSkew,
+		});
 
 		// if successful, memoize the token
 		memoizedVerifyTokens.set(idToken.claims.jti, true);
@@ -1104,7 +787,7 @@ export class Token<
 		idTokenRaw,
 		idTokenClockSkew,
 		nonce,
-	}: Omit<DecodeTokenParams, 'options'>) {
+	}: Omit<DecodeTokensParams, 'options'>) {
 		return decodeTokens<AC, IC>({
 			accessTokenRaw,
 			accessTokenClockSkew,
