+++ libs/@guardian/identity-auth/src/tokenVerifier.ts
@@ -1,6 +1,17 @@
 import type { OAuthUrls, ProfileUrl } from './@types/OAuth';
+import type {
+	AccessTokenClaims,
+	CustomClaims,
+	IDTokenClaims,
+	JWTObject,
+} from './@types/Token';
 import { OAuthError } from './error';
-import { decodeToken, verifySignature } from './token';
+import {
+	decodeToken,
+	genericVerifyAccessTokenClaims,
+	genericVerifyIdTokenClaims,
+	verifySignature,
+} from './lib/token';
 
 interface TokenVerifierOptions {
 	issuer: `${ProfileUrl}/oauth2/${string}`;
@@ -30,6 +41,13 @@ export class TokenVerifier {
 			userinfoUrl: `${this.#options.issuer}/v1/userinfo`,
 		};
 	}
+	async #decodeAndVerifyToken<T extends CustomClaims = CustomClaims>(
+		token: string,
+	): Promise<JWTObject<T>> {
+		const decodedToken = decodeToken<T>(token);
+		await verifySignature(this.#oauthUrls.keysUrl, decodedToken, token);
+		return decodedToken;
+	}
 
 	/**
 	 * Verify an OAuth access token
@@ -42,94 +60,25 @@ export class TokenVerifier {
 	 * - The cid claim matches the client ID the verifier is constructed with.
 	 * - Any custom claim assertions that you add are confirmed
 	 */
-	public async verifyAccessToken({
+	public async verifyAccessToken<T extends CustomClaims = CustomClaims>({
 		accessToken,
 		audience,
 		scopes,
-	}: AccessTokenVerifierParams) {
-		const decodedToken = await this.#verifyToken(accessToken);
+	}: AccessTokenVerifierParams): Promise<JWTObject<AccessTokenClaims<T>>> {
+		const decodedToken =
+			await this.#decodeAndVerifyToken<AccessTokenClaims<T>>(accessToken);
 
-		// get the local time in seconds
-		const localTime = Math.floor(Date.now() / 1000);
-
-		// check that the iat (issued at) and exp (expiry) decodedToken.payload are present
-		if (!decodedToken.payload.iat || !decodedToken.payload.exp) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Missing iat or exp claim in access token',
-				message: 'Token does not contain required decodedToken.payload',
-			});
-		}
-
-		// check that the iat isn't after the exp
-		if (decodedToken.payload.iat > decodedToken.payload.exp) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'iat claim is after exp claim in access token',
-				message: 'Token has expired before it was issued',
-			});
-		}
-
-		// check the token hasn't expired
-		if (localTime > decodedToken.payload.exp) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token has expired',
-				message: 'Token has expired',
-			});
-		}
-
-		// check the token wasn't issued in the future
-		if (decodedToken.payload.iat > localTime) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token was issued in the future',
-				message: 'Token was issued in the future',
-			});
-		}
-
-		// check audience claim
-		if (decodedToken.payload.aud !== audience) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token audience does not match expected audience',
-				message: 'Token audience does not match expected audience',
-			});
-		}
-
-		// check issuer claim
-		if (decodedToken.payload.iss !== this.#options.issuer) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token issuer does not match expected issuer',
-				message: 'Token issuer does not match expected issuer',
-			});
-		}
-
-		// check client ID claim (if it exists)
-		if (
-			this.#options.clientId &&
-			decodedToken.payload.cid !== this.#options.clientId
-		) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token client ID does not match expected client ID',
-				message: 'Token client ID does not match expected client ID',
-			});
-		}
-
-		if (!decodedToken.payload.scp || decodedToken.payload.scp.length === 0) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token is missing required scopes',
-				message: 'Token is missing required scopes',
-			});
-		}
+		genericVerifyAccessTokenClaims({
+			claims: decodedToken.payload,
+			audience,
+			issuer: this.#options.issuer,
+			clientId: this.#options.clientId,
+		});
 
 		// Check that all scopes in scopes array (if it exists) are present in the token
 		if (scopes && scopes.length > 0) {
-			const everyScopeExists = scopes.every(
-				(scope) => decodedToken.payload.scp?.includes(scope),
+			const everyScopeExists = scopes.every((scope) =>
+				decodedToken.payload.scp.includes(scope),
 			);
 
 			if (!everyScopeExists) {
@@ -177,8 +126,14 @@ export class TokenVerifier {
 				}
 			}
 		}
+
+		return decodedToken;
 	}
-	public async verifyIdToken({ idToken, nonce }: IdTokenVerifierParams) {
+
+	public async verifyIdToken<T extends CustomClaims = CustomClaims>({
+		idToken,
+		nonce,
+	}: IdTokenVerifierParams): Promise<JWTObject<IDTokenClaims<T>>> {
 		if (!this.#options.clientId) {
 			throw new OAuthError({
 				error: 'invalid_token',
@@ -187,61 +142,16 @@ export class TokenVerifier {
 			});
 		}
 
-		const decodedToken = await this.#verifyToken(idToken);
+		const decodedToken =
+			await this.#decodeAndVerifyToken<IDTokenClaims<T>>(idToken);
 
-		// get the local time in seconds
-		const localTime = Math.floor(Date.now() / 1000);
+		genericVerifyIdTokenClaims({
+			claims: decodedToken.payload,
+			nonce,
+			clientId: this.#options.clientId,
+			issuer: this.#options.issuer,
+		});
 
-		// aud claim matches clientId
-		if (decodedToken.payload.aud !== this.#options.clientId) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description:
-					'Token `aud` claim does not match expected `clientId`',
-				message: 'Token `aud` claim does not match expected `clientId`',
-			});
-		}
-
-		// iss claim matches issuer
-		if (decodedToken.payload.iss !== this.#options.issuer) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token issuer does not match expected issuer',
-				message: 'Token issuer does not match expected issuer',
-			});
-		}
-
-		// nonce claim matches nonce (if it exists)
-		if (nonce && decodedToken.payload.nonce !== nonce) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token nonce does not match expected nonce',
-				message: 'Token nonce does not match expected nonce',
-			});
-		}
-
-		// check that the iat (issued at) claim is present
-		if (!decodedToken.payload.iat) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Missing iat claim in ID token',
-				message: 'Token does not contain required claims',
-			});
-		}
-
-		// check the token wasn't issued in the future
-		if (decodedToken.payload.iat > localTime) {
-			throw new OAuthError({
-				error: 'invalid_token',
-				error_description: 'Token was issued in the future',
-				message: 'Token was issued in the future',
-			});
-		}
-	}
-
-	async #verifyToken(token: string) {
-		const decodedToken = decodeToken(token);
-		await verifySignature(this.#oauthUrls.keysUrl, decodedToken, token);
 		return decodedToken;
 	}
 }
